% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cutoff_at_first_empty_bin.R
\name{cluster_cutoff_at_first_empty_bin}
\alias{cluster_cutoff_at_first_empty_bin}
\title{cluster_cutoff_at_first_empty_bin function}
\usage{
cluster_cutoff_at_first_empty_bin(heights, diam, num_bins_when_clustering)
}
\arguments{
\item{heights}{Height values in hierarchical clustering.}

\item{diam}{Maximum distance between points in a level set.}

\item{num_bins_when_clustering}{Controls how many bins there are in the histogram used to determine cutoff. values}
}
\value{
Numerical value for cutoff point of hierarchical cluster diagram.
}
\description{
This function decides where to cut the hierarchical clustering tree to define clusters within a level set.
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapper1D}}, \code{\link{mapper2D}}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
