% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\docType{class}
\name{Cover-class}
\alias{Cover-class}
\alias{cover_ref}
\title{Reference Class (R5) implementation of Mapper for the TDAmapper package in R
Cover reference class
filter_values := Filter values 
num_intervals := vector of number of bins to cover the Z with (per dimension)
percent_overlap := vector of overlap percentages
level_sets := list of level_sets (including min/max bounds)}
\description{
Reference Class (R5) implementation of Mapper for the TDAmapper package in R
Cover reference class
filter_values := Filter values 
num_intervals := vector of number of bins to cover the Z with (per dimension)
percent_overlap := vector of overlap percentages
level_sets := list of level_sets (including min/max bounds)
}

