% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperEdges.R
\name{mapperEdges}
\alias{mapperEdges}
\title{mapperEdges function}
\usage{
mapperEdges(m)
}
\arguments{
\item{m}{An object of class TDAmapper that is the output of the mapper function.}
}
\value{
A data frame describing the edges in the graph of the mapper output.
}
\description{
The input to this function is a TDAmapper class object and the output
is a data frame of edges that can be used as input to the networkD3 
plot utility.
}
\examples{
\dontrun{
X <- data.frame( x=2*cos(0.5*(1:100)), y=sin(1:100) )
f <- X$y
m1 <- mapper(
       distance_matrix = dist(X),
       filter_values = f,
       num_intervals = 10,
       percent_overlap = 50,
       num_bins_when_clustering = 10)
       
pt_labels <- 1:length(f)
vertices <- mapperVertices(m1, pt_labels)
edges <- mapperEdges(m1)

# interactive plot
forceNetwork(Nodes = nodes, Links = links, 
            Source = "Linksource", Target = "Linktarget",
            Value = "Linkvalue", NodeID = "Nodename",
            Group = "Nodegroup", opacity = 0.8, 
            linkDistance = 10, charge = -400)
}
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapperVertices}}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
\keyword{mapperEdges}
