% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapperRef.R
\docType{class}
\name{MapperRef-class}
\alias{MapperRef-class}
\alias{mapper_ref}
\title{Reference Class (R5) implementation of Mapper for the TDAmapper package in R
Composes a set of classes to perform Mapper efficiently.
Author: Matt Piekenbrock}
\description{
Reference Class (R5) implementation of Mapper for the TDAmapper package in R
Composes a set of classes to perform Mapper efficiently.
Author: Matt Piekenbrock
}

