% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.TDAmapper.R
\name{plot.TDAmapper}
\alias{plot.TDAmapper}
\title{Plot an object of class TDAmapper using igraph}
\usage{
\method{plot}{TDAmapper}(x, ...)
}
\arguments{
\item{x}{An object of class "TDAmapper".}

\item{...}{Optional arguments to the igraph plot() function.}
}
\value{
plot of the mapper object using igraph
}
\description{
Plot method defined for objects of class "TDAmapper" using auto layout for igraph.
}
\examples{
\dontrun{
plot(x, ...)
}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapper1D}}, \code{\link{mapper2D}}
}
\keyword{plot.TDAmapper}

