% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.TDAmapper.R
\name{print.TDAmapper}
\alias{print.TDAmapper}
\title{Print an object of class TDAmapper}
\usage{
\method{print}{TDAmapper}(x, ...)
}
\arguments{
\item{x}{An object of class "TDAmapper".}

\item{...}{Optional arguments.}
}
\value{
Print the number of vertices in the mapper object
}
\description{
Print method defined for objects of class "TDAmapper".
}
\examples{
\dontrun{
print(x, ...)
}
}
\author{
Paul Pearson, \email{pearsonp@hope.edu}
}
\references{
\url{https://github.com/paultpearson/TDAmapper}
}
\seealso{
\code{\link{mapper1D}}, \code{\link{mapper2D}}
}
\keyword{print.TDAmapper}

